+++ slugathon/game/Action.py
@@ -88,8 +88,8 @@ class FormGame(Action):
         start_time: float,
         min_players: int,
         max_players: int,
-        ai_time_limit: int,
-        player_time_limit: int,
+        ai_time_limit: float,
+        player_time_limit: float,
         player_class: str,
         player_info: str,
     ):
